refactor: LottoNumbers 클래스 util package로 이동
