+++ src/test/java/dev/dahye/lotto/util/LottoNumbersTest.java
@@ -1,4 +1,4 @@
-package dev.dahye.lotto.domain;
+package dev.dahye.lotto.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
